requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
uvloop==0.20.0; sys_platform != 'win32'
fastapi==0.115.4
uvicorn==0.32.0
jinja2==3.1.4
python-multipart==0.0.17
passlib[bcrypt]==1.7.4
